wxGUI/gmodeler: refactor - create model_items.py
move individual model items from model.py -> model_items.py
